arreglar controllers y añadir service
se modifico, arreglo y añadio controllers
- UsuarioController
- CuentaBancariaController
Tambien se añadio Services :
- UsuariosService
-UsuariosServiceImpl

+++ src/main/java/banco/dao/RolUsuarioDAO.java
@@ -6,8 +6,8 @@ import java.util.Optional;
 
 public interface RolUsuarioDAO {
     List<RolUsuario> findAll();
-    Optional<RolUsuario> findById(Long id);
+    Optional<RolUsuario> findById(Integer id);
     RolUsuario save(RolUsuario rolUsuario);
     RolUsuario update(RolUsuario rolUsuario);
-    void deleteById(Long id);
+    void deleteById(Integer id);
 }
